Moved tests to own dir, and used Pathlib
